--- app/src/main/java/com/example/donotlate/feature/directionRoute/data/DirectionsRepository.kt
@@ -1,7 +1,0 @@
-package com.example.donotlate.feature.directionRoute.data
-
-import com.example.donotlate.feature.directionRoute.domain.DirectionsEntity
-
-interface DirectionsRepository {
-    suspend fun getDirections(origin: String, destination: String, mode: String): DirectionsEntity
-}
