+++ app/src/main/java/com/example/donotlate/feature/directionRoute/domain/DirectionsRepository.kt
@@ -1,0 +1,5 @@
+package com.example.donotlate.feature.directionRoute.domain
+
+interface DirectionsRepository {
+    suspend fun getDirections(origin: String, destination: String, mode: String): DirectionsEntity
+}
