+++ app/src/main/java/com/example/donotlate/feature/main/presentation/view/MainFragment.kt
@@ -41,7 +41,6 @@ class MainFragment : Fragment() {
             appContainer.getCurrentUserUseCase,
             appContainer.imageUploadUseCase,
             appContainer.firebaseAuth
-//
         )
     }
